requirements/base.txt
Django==3.1.2
psycopg2==2.8.6
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
suds-community==0.8.4
requests==2.24.0
urllib3==1.25.11
